Added machine-mode flag and can now output metrics.
Changed default ouput to be non-machine mode (user mode), specifically,
now only outputs the program's return as exit code in machine mode.
In user mode, we print the full details of the returned operands and
metrics by default. The exit code should only in this mode be used
to signify if the simulation succeeded (whatever that means)